Wrap the method call with try catch block
